+++ smc/src/main/java/com/tlcsdm/smc/codeDev/ecm/U2AEcmScript.java
@@ -38,7 +38,7 @@ import javafx.stage.Stage;
  * @author: unknowIfGuestInDream
  * @date: 2023/3/26 21:17
  */
-public class U2AEcmScript extends AbstractEcmScript {
+public class U2AEcmScript extends AbstractU2XFamilyScript {
 
     @Override
     protected void initDefaultValue() {
